Android 6.0.0 release 1

Conflicts:
	include/media/stagefright/ACodec.h
	media/libmedia/AudioTrack.cpp
	media/libmedia/IMediaHTTPConnection.cpp
	media/libmediaplayerservice/nuplayer/GenericSource.cpp
	media/libmediaplayerservice/nuplayer/NuPlayer.cpp
	media/libstagefright/ACodec.cpp
	media/libstagefright/Android.mk
	media/libstagefright/MPEG4Extractor.cpp
	media/libstagefright/MediaCodec.cpp
	media/libstagefright/SampleTable.cpp
	media/libstagefright/codecs/on2/dec/SoftVPX.cpp
	media/libstagefright/colorconversion/SoftwareRenderer.cpp
	media/libstagefright/include/OMXNodeInstance.h
	media/libstagefright/omx/OMXNodeInstance.cpp
	services/audiopolicy/managerdefault/AudioPolicyManager.cpp

+++ media/libstagefright/SampleTable.cpp
@@ -235,7 +235,7 @@ status_t SampleTable::setSampleToChunkParams(
         return ERROR_MALFORMED;
     }
 
-    if (SIZE_MAX / sizeof(SampleToChunkEntry) <= mNumSampleToChunkOffsets)
+    if (SIZE_MAX / sizeof(SampleToChunkEntry) <= (size_t)mNumSampleToChunkOffsets)
         return ERROR_OUT_OF_RANGE;
 
     mSampleToChunkEntries =
@@ -344,7 +344,7 @@ status_t SampleTable::setTimeToSampleParams(
 
     mTimeToSampleCount = U32_AT(&header[4]);
     uint64_t allocSize = (uint64_t)mTimeToSampleCount * 2 * sizeof(uint32_t);
-    if (allocSize > SIZE_MAX) {
+    if (allocSize > UINT32_MAX) {
         return ERROR_OUT_OF_RANGE;
     }
     mTimeToSample = new (std::nothrow) uint32_t[mTimeToSampleCount * 2];
@@ -392,7 +392,7 @@ status_t SampleTable::setCompositionTimeToSampleParams(
 
     mNumCompositionTimeDeltaEntries = numEntries;
     uint64_t allocSize = (uint64_t)numEntries * 2 * sizeof(uint32_t);
-    if (allocSize > SIZE_MAX) {
+    if (allocSize > UINT32_MAX) {
         return ERROR_OUT_OF_RANGE;
     }
